+++ CHANGES.md
@@ -5,7 +5,7 @@ Changelog
 1.0.0a8 (unreleased)
 --------------------
 
-- Nothing changed yet.
+- Use a collective.taxonomy to manage the available section CSS classes. [sverbois]
 
 
 1.0.0a7 (2021-12-22)